Intersect each ray with the scene's bounding box to obtain near and far values. The bounding box is taken directly from colmap. Additional workarounds for each view having its own camera parameters. Saving and loading from JSON View and NeRFDatasetParams. Simplified NeRF by removing an unnecessary coarse network. Attempt to add antialiasing, representing the ray as a diverging cone (based on the angular dimensions of a camera pixel with given parameters), rather than an infinitely thin one. Removed some warnings. A minor improvement to numerical stability: normalization with constraints; when calculating transparency, we add logarithms instead of multiplying exponentials. Fixed a bug where large scenes would freeze in extreme quality. Undistortion added.

+++ src/CuSHEncoder.h
@@ -14,7 +14,7 @@ public:
 		const std::string &module_name,
 		const int input_dim = 3,
 		const int degree = 4
-	) : BaseEmbedderImpl(module_name), InputDim(input_dim), Degree(degree), OutputDims(pow(degree, 2))
+	) : BaseEmbedderImpl(module_name), InputDim(input_dim), Degree(degree), OutputDims((int)powf((float)degree, 2))
 	{
 		//assert input_dim == 3
 		//assert degree >= 1 && self.degree <= 5
